apache-airflow/requirements.txt
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark 
apache-airflow-providers-amazon
pyspark==3.5.3